feat: 수익을 관리하는 Profits 추가
