+++ src/main/java/domain/betting/Bettings.java
@@ -7,13 +7,9 @@ import java.util.Map;
 
 public class Bettings {
 
-    private static final int PREFIX_INITIAL_PROFIT = 0;
-
     private final Map<PlayerName, BettingMoney> bettingMoneyByPlayer;
-    private final Map<PlayerName, Profit> profitByPlayer;
 
     public Bettings() {
-        this.profitByPlayer = new HashMap<>();
         this.bettingMoneyByPlayer = new HashMap<>();
     }
 
@@ -23,22 +19,6 @@ public class Bettings {
 
     public BettingMoney getPlayerBettingMoney(PlayerName playerName) {
         return bettingMoneyByPlayer.get(playerName);
-    }
-
-    public void settleBettingMoney(PlayerName playerName, double bettingRate) {
-        BettingMoney bettingMoney = bettingMoneyByPlayer.get(playerName);
-        profitByPlayer.put(playerName, bettingMoney.withRate(bettingRate));
-    }
-
-    public Profit calculateDealerProfit() {
-        return profitByPlayer.values()
-                .stream()
-                .reduce(new Profit(PREFIX_INITIAL_PROFIT), Profit::addProfit)
-                .reverseProfit();
-    }
-
-    public Profit getPlayerProfit(PlayerName playerName) {
-        return profitByPlayer.get(playerName);
     }
 
 }
